ecc: add test for config
