+++ .github/workflows/ecc.yml
@@ -32,7 +32,7 @@ jobs:
       run: |
           sudo make -C eunomia-cc install-deps
           cargo install clippy-sarif sarif-fmt grcov
-	        rustup component add llvm-tools-preview
+          rustup component add llvm-tools-preview
 
     - name: Install Rust toolchain
       uses: actions-rs/toolchain@16499b5e05bf2e26879000db0c1d13f7e13fa3af #@v1
